+++ AdapterUnitTests/MaioAdapterTests/AUTMaioAdapterTests.m
@@ -1,5 +1,7 @@
 #import "GADMediationAdapterMaio.h"
 
+#import "GADMMaioConstants.h"
+
 #import <AdapterUnitTestKit/AUTKAdConfiguration.h>
 #import <AdapterUnitTestKit/AUTKAdapterSetUpAssertions.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
@@ -12,6 +14,18 @@
 @end
 
 @implementation AUTMaioAdapterTests
+
+- (void)setUp {
+  [super setUp];
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
+}
+
+- (void)tearDown {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
+  [super tearDown];
+}
 
 - (void)testAdapterVersion {
   GADVersionNumber version = [GADMediationAdapterMaio adapterVersion];
@@ -35,9 +49,54 @@
   XCTAssertLessThanOrEqual(version.patchVersion, 99);
 }
 
-- (void)testSetUp {
+- (void)testSetUpSucceeds {
   AUTKMediationServerConfiguration *config = [[AUTKMediationServerConfiguration alloc] init];
   AUTKWaitAndAssertAdapterSetUpWithConfiguration([GADMediationAdapterMaio class], config);
+}
+
+- (void)testSetUpSucceedsWhenTagForChildDirectedTreatmentIsFalse {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @NO;
+
+  AUTKMediationServerConfiguration *config = [[AUTKMediationServerConfiguration alloc] init];
+  AUTKWaitAndAssertAdapterSetUpWithConfiguration([GADMediationAdapterMaio class], config);
+}
+
+- (void)testSetUpSucceedsWhenTagForUnderAgeOfConsentIsFalse {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @NO;
+
+  AUTKMediationServerConfiguration *config = [[AUTKMediationServerConfiguration alloc] init];
+  AUTKWaitAndAssertAdapterSetUpWithConfiguration([GADMediationAdapterMaio class], config);
+}
+
+- (void)testSetUpFailsWhenTagForChildTreatmentIsTrue {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @YES;
+
+  NSString *errorDescription = @"The request had age-restricted treatment, but maio SDK "
+                               @"cannot receive age-restricted signals.";
+  NSDictionary *errorUserInfo = @{
+    NSLocalizedDescriptionKey : errorDescription,
+    NSLocalizedFailureReasonErrorKey : errorDescription
+  };
+  NSError *error = [[NSError alloc] initWithDomain:GADMMaioErrorDomain
+                                              code:GADMAdapterMaioErrorChildUser
+                                          userInfo:errorUserInfo];
+  AUTKWaitAndAssertAdapterSetUpFailureWithConfiguration(
+      [GADMediationAdapterMaio class], [[AUTKMediationServerConfiguration alloc] init], error);
+}
+- (void)testSetUpFailsWhenTagForUnderAgeOfConsentIsTrue {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @YES;
+
+  NSString *errorDescription = @"The request had age-restricted treatment, but maio SDK "
+                               @"cannot receive age-restricted signals.";
+  NSDictionary *errorUserInfo = @{
+    NSLocalizedDescriptionKey : errorDescription,
+    NSLocalizedFailureReasonErrorKey : errorDescription
+  };
+  NSError *error = [[NSError alloc] initWithDomain:GADMMaioErrorDomain
+                                              code:GADMAdapterMaioErrorChildUser
+                                          userInfo:errorUserInfo];
+  AUTKWaitAndAssertAdapterSetUpFailureWithConfiguration(
+      [GADMediationAdapterMaio class], [[AUTKMediationServerConfiguration alloc] init], error);
 }
 
 - (void)testNetworkExtras {