+++ AdapterUnitTests/MaioAdapterTests/AUTMaioBannerAdTests.m
@@ -26,36 +26,41 @@ static const NSInteger kMaioUnknownFailureErroCode = 99999;
 }
 
 - (void)setUp {
+  [super setUp];
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
   _adapter = [[GADMediationAdapterMaio alloc] init];
   _bannerMock = OCMClassMock([MaioBannerView class]);
   OCMStub([_bannerMock alloc]).andReturn(_bannerMock);
 }
 
 - (void)tearDown {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
   OCMVerifyAll(_bannerMock);
+  [super tearDown];
 }
 
 - (AUTKMediationBannerAdEventDelegate *)loadBannerAd {
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{GADMMaioAdapterZoneIdKey : @"zoneID"};
-  AUTKMediationBannerAdConfiguration *config =
-      [[AUTKMediationBannerAdConfiguration alloc] init];
+  AUTKMediationBannerAdConfiguration *config = [[AUTKMediationBannerAdConfiguration alloc] init];
   config.credentials = credentials;
   config.adSize = GADAdSizeBanner;
   config.isTestRequest = YES;
 
-  OCMExpect([_bannerMock initWithZoneId:@"zoneID" size:MaioBannerSize.banner]).andReturn(_bannerMock);
+  OCMExpect([_bannerMock initWithZoneId:@"zoneID" size:MaioBannerSize.banner])
+      .andReturn(_bannerMock);
   OCMStub([_bannerMock setDelegate:[OCMArg checkWithBlock:^BOOL(id obj) {
-    if ([obj conformsToProtocol:@protocol(MaioBannerDelegate)]) {
-      id<MaioBannerDelegate> maioDelegate = obj;
-      [maioDelegate didLoad:self->_bannerMock];
-      return YES;
-    }
-    return NO;
-  }]]);
+                         if ([obj conformsToProtocol:@protocol(MaioBannerDelegate)]) {
+                           id<MaioBannerDelegate> maioDelegate = obj;
+                           [maioDelegate didLoad:self->_bannerMock];
+                           return YES;
+                         }
+                         return NO;
+                       }]]);
 
-  AUTKMediationBannerAdEventDelegate *delegate =
-      AUTKWaitAndAssertLoadBannerAd(_adapter, config);
+  AUTKMediationBannerAdEventDelegate *delegate = AUTKWaitAndAssertLoadBannerAd(_adapter, config);
   XCTAssertNotNil(delegate);
 
   return delegate;
@@ -64,8 +69,7 @@ static const NSInteger kMaioUnknownFailureErroCode = 99999;
 - (void)loadBannerAdFailureWithInvalidAdSize {
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{GADMMaioAdapterZoneIdKey : @"zoneID"};
-  AUTKMediationBannerAdConfiguration *config =
-    [[AUTKMediationBannerAdConfiguration alloc] init];
+  AUTKMediationBannerAdConfiguration *config = [[AUTKMediationBannerAdConfiguration alloc] init];
   config.credentials = credentials;
   config.isTestRequest = YES;
 
@@ -78,21 +82,21 @@ static const NSInteger kMaioUnknownFailureErroCode = 99999;
 - (void)loadBannerAdFailureWithErrorCode:(NSInteger)errorCode {
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{GADMMaioAdapterZoneIdKey : @"zoneID"};
-  AUTKMediationBannerAdConfiguration *config =
-      [[AUTKMediationBannerAdConfiguration alloc] init];
+  AUTKMediationBannerAdConfiguration *config = [[AUTKMediationBannerAdConfiguration alloc] init];
   config.credentials = credentials;
   config.adSize = GADAdSizeBanner;
   config.isTestRequest = YES;
 
-  OCMExpect([_bannerMock initWithZoneId:@"zoneID" size:MaioBannerSize.banner]).andReturn(_bannerMock);
+  OCMExpect([_bannerMock initWithZoneId:@"zoneID" size:MaioBannerSize.banner])
+      .andReturn(_bannerMock);
   OCMStub([_bannerMock setDelegate:[OCMArg checkWithBlock:^BOOL(id obj) {
-    if ([obj conformsToProtocol:@protocol(MaioBannerDelegate)]) {
-      id<MaioBannerDelegate> maioDelegate = obj;
-      [maioDelegate didFailToLoad:self->_bannerMock errorCode:errorCode];
-      return YES;
-    }
-    return NO;
-  }]]);
+                         if ([obj conformsToProtocol:@protocol(MaioBannerDelegate)]) {
+                           id<MaioBannerDelegate> maioDelegate = obj;
+                           [maioDelegate didFailToLoad:self->_bannerMock errorCode:errorCode];
+                           return YES;
+                         }
+                         return NO;
+                       }]]);
 
   NSError *expectedError = [[NSError alloc] initWithDomain:GADMMaioSDKErrorDomain
                                                       code:errorCode
@@ -100,12 +104,53 @@ static const NSInteger kMaioUnknownFailureErroCode = 99999;
   AUTKWaitAndAssertLoadBannerAdFailure(_adapter, config, expectedError);
 }
 
-- (void)testLoadBannerAd {
+- (void)loadBannerAdFailureForChildUser {
+  AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
+  credentials.settings = @{GADMMaioAdapterZoneIdKey : @"zoneID"};
+  AUTKMediationBannerAdConfiguration *config = [[AUTKMediationBannerAdConfiguration alloc] init];
+  config.credentials = credentials;
+  config.adSize = GADAdSizeBanner;
+  config.isTestRequest = YES;
+
+  NSString *errorDescription = @"The request had age-restricted treatment, but maio SDK "
+                               @"cannot receive age-restricted signals.";
+  NSDictionary *errorUserInfo = @{
+    NSLocalizedDescriptionKey : errorDescription,
+    NSLocalizedFailureReasonErrorKey : errorDescription
+  };
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMMaioErrorDomain
+                                                      code:GADMAdapterMaioErrorChildUser
+                                                  userInfo:errorUserInfo];
+
+  AUTKWaitAndAssertLoadBannerAdFailure(_adapter, config, expectedError);
+}
+
+- (void)testLoadBannerAdSucceeds {
+  [self loadBannerAd];
+}
+
+- (void)testLoadBannerAdSucceedsWhenTagForChildDirectedTreatmentIsFalse {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @NO;
+  [self loadBannerAd];
+}
+
+- (void)testLoadBannerAdSucceedsWhenTagForUnderAgeOfConsentIsFalse {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @NO;
   [self loadBannerAd];
 }
 
 - (void)testLoadFailureWithInvalidAdSize {
   [self loadBannerAdFailureWithInvalidAdSize];
+}
+
+- (void)testLoadBannerAdFailsWhenTagForChildDirectedTreatmentIsTrue {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @YES;
+  [self loadBannerAdFailureForChildUser];
+}
+
+- (void)testLoadBannerAdFailsWhenTagForUnderAgeOfConsentIsTrue {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @YES;
+  [self loadBannerAdFailureForChildUser];
 }
 
 - (void)testMaioAdLoadFailure {